// core/apps/audit/src/modules/vendor-position/helper/enums.ts
export enum PositionDataEnum {
  ID = 'id',
  NAME = 'name',
  TOTAL_ORDER = 'totalOrder',
  CHECKED_IN = 'totalCheckedInStaff',
  CHECKED_OUT = 'checkedOut',
  CURRENT_CHECKED_IN = 'currentCheckedInStaff',
  CHECKED_OUT_PERCENTAGE = 'totalCheckedOutPercentage',
  CHECKED_IN_PERCENTAGE = 'totalCheckedInPercentage',
}